requirements.txt
pyspark==3.5.1
kafka-python==2.0.2
python-binance==1.0.19
pandas==2.2.2
numpy==1.26.4
pyarrow==15.0.2
fastparquet==2024.2.0
python-dotenv==1.0.1
pydantic==2.7.1
loguru==0.7.2
tenacity==8.3.0
pytest==8.2.0
black==24.4.2
isort==5.13.2